capemgr: Priority on capemgr.enable_partno option
When booting from a device that's providing the rootfs via a
cape fragment, you face the following problem as it appears
on the beaglebone black.

The priority of the emmc cape is low (positive number) so that
any cape that is detected and uses the resources of the emmc
interface can work (for example the camera cape). This means that
we have to be booting from external sd card.

However if there is no conflict, and the cape dtbo is not included
in the kernel (via the firmware builtin configure option) the capemgr
will attempt to load the cape too early, at a time that the rootfs
cape hasn't been loaded yet.

In order to fix that an optional third parameter is added to
the specification of a cape in the eanble_partno option which allow
you to specify the priority.

so capemgr.enable_partno=[PART[:REV[:PRIO]]][,PART..]

To load the cape fragment after all the builtin capes have loaded
use a priority higher than 10; i.e. to load cape foo use:

capemgr.enable_partno=foo:00A0:10

Signed-off-by: Pantelis Antoniou <panto@antoniou-consulting.com>